+++ cyclope/apps/forum/models.py
@@ -35,8 +35,11 @@ from cyclope.models import BaseContent
 
 
 class Topic(BaseContent, Collectible):
-    author = models.ForeignKey(User, verbose_name=_('author'))
     text = models.TextField(_('text'))
+    
+    @property
+    def author(self):
+        return self.user
 
     class Meta:
         verbose_name = _('topic')